backend/phase-2-web/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "evolution-todo-phase-2"
version = "2.0.0"
description = "Evolution of Todo - Phase 2: Full-Stack Web Application"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Development Team", email = "dev@example.com"}
]

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.36",
    "sqlmodel==0.0.22",
    "psycopg[binary]==3.2.1",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "alembic==1.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "mypy>=1.14.1",
    "ruff>=0.9.2",
    "black>=24.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = []

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"
